create skilltree object
