+++ api/skilltree/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from .models import SkillTree
+
+# Register your models here.
+
+admin.site.register(SkillTree)
