+++ api/skilltree/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class SkilltreeConfig(AppConfig):
+    name = 'skilltree'
