+++ api/skilltree/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from .models import SkillTree
+
+class SkillTreeSerializer(serializers.HyperlinkedModelSerializer):
+
+    class Meta:
+        model = SkillTree
+        fields = ('id', 'tree_name', 'tree_desc')
