--- apireact.code-workspace
@@ -1,11 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "react/src"
-		},
-		{
-			"path": "api"
-		}
-	],
-	"settings": {}
-}
